# Embed mutex in the server's websocket client