@tigrannajaryan 2023-09-19
Hint: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat is very useful for showing before/after benchmark comparison.

Here is what it produces for your benchmark results:
```
name             old time/op  new time/op  delta
SendToClient-24   238µs ±50%   262µs ±33%   ~     (p=0.690 n=5+5)
```
Looks like it is close, but could use a more stable execution environment for benchmarking. :-)